@@ -1,29 +1,23 @@
-package com.liyu.huahui;
+package com.liyu.huahui
 
-import android.app.Application;
-import android.content.Context;
+import android.app.Application
+import android.content.SharedPreferences
+import com.liulishuo.filedownloader.FileDownloader
+import com.liyu.huahui.model.Word
+import org.litepal.LitePal
 
-import com.liulishuo.filedownloader.FileDownloader;
-
-import org.litepal.LitePal;
-
-/**
- * Created by liyu on 2017/3/7.
- */
-
-public class App extends Application {
-
-    private static Context mContext;
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        mContext = getApplicationContext();
-        LitePal.initialize(this);
-        FileDownloader.setupOnApplicationOnCreate(this);
+class App : Application() {
+    private lateinit var sharedPreferences: SharedPreferences
+    override fun onCreate() {
+        super.onCreate()
+        LitePal.initialize(this)
+        FileDownloader.setupOnApplicationOnCreate(this)
+        sharedPreferences = getSharedPreferences("data", MODE_PRIVATE)
+        accent = Word.Accent.values()[sharedPreferences.getInt("accent", Word.Accent.US.id)-1]
     }
 
-    public static Context getContext(){
-        return mContext;
+    companion object {
+        @JvmStatic
+        var accent: Word.Accent = Word.Accent.US
     }
 }
