@@ -1,90 +1,50 @@
-package com.liyu.huahui.model;
+package com.liyu.huahui.model
 
-import org.litepal.crud.DataSupport;
-
-import java.io.Serializable;
+import org.litepal.crud.DataSupport
+import java.io.Serializable
 
 /**
  * Created by liyu on 2017/3/2.
  */
+class Word : DataSupport(), Serializable {
+    var name: String? = null
+    var correctPhonetic: String? = null
+    var wrongPhonetic: String? = null
 
-public class Word extends DataSupport implements Serializable {
+    //+"&type="+accent;
+    val voiceUrl: String?
+        get() = "http://dict.youdao.com/dictvoice?audio=$name" //+"&type="+accent;
+    private var sourceFrom = 0
+        private set
+    private var accent = 0
+        private set
 
-    private String name;
-    private String correctPhonetic;
-    private String wrongPhonetic;
-    private String voiceUrl;
-    private int sourceFrom;
-
-    public int getSourceFrom() {
-        return sourceFrom;
+    fun setSourceFrom(sourceFrom: From) {
+        this.sourceFrom = sourceFrom.from
     }
 
-    public void setSourceFrom(From sourceFrom) {
-        this.sourceFrom = sourceFrom.getFrom();
+    override fun equals(o: Any?): Boolean {
+        if (this === o) return true
+        if (o == null || javaClass != o.javaClass) return false
+        val word = o as Word
+        return if (name != null) name!!.toLowerCase() == word.name!!.toLowerCase() else word.name == null
     }
 
-    public String getVoiceUrl() {
-        return voiceUrl;
+    override fun hashCode(): Int {
+        var result = name?.hashCode() ?: 0
+        result = 31 * result + (correctPhonetic?.hashCode() ?: 0)
+        result = 31 * result + (wrongPhonetic?.hashCode() ?: 0)
+        result = 31 * result + sourceFrom
+        result = 31 * result + accent
+        return result
     }
 
-    public void setVoiceUrl(String voiceUrl) {
-        this.voiceUrl = voiceUrl;
-    }
-
-    public String getCorrectPhonetic() {
-        return correctPhonetic;
-    }
-
-    public void setCorrectPhonetic(String correctPhonetic) {
-        this.correctPhonetic = correctPhonetic;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getWrongPhonetic() {
-        return wrongPhonetic;
-    }
-
-    public void setWrongPhonetic(String wrongPhonetic) {
-        this.wrongPhonetic = wrongPhonetic;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        Word word = (Word) o;
-
-        return name != null ? name.toLowerCase().equals(word.name.toLowerCase()) : word.name == null;
+    enum class From(var from: Int) {
+        LOCAL(1), NETWORK(2);
 
     }
 
-    public enum From {
-        LOCAL(1),
-        NETWORK(2);
-
-        private int from;
-
-        From(int from) {
-            this.from = from;
-        }
-
-        public int getFrom() {
-            return from;
-        }
-
-        public void setFrom(int from) {
-            this.from = from;
-        }
+    enum class Accent(var id: Int) {
+        GB(1), US(2);
     }
-
-
 }
